[REFACTOR] Reorganizng the directory
