+++ package.json
@@ -7,7 +7,7 @@
     "node": ">=12.0.0"
   },
   "scripts": {
-    "server": "node src/server",
+    "server": "node src/server/server",
     "test": "cross-env NODE_ENV=production xo && nyc --reporter=html --reporter=text ava --verbose --timeout=1m",
     "test-no": "cross-env NODE_ENV=development ava tests/*.js --verbose --timeout=1m",
     "coverage": "nyc report --reporter=text-lcov | coveralls"
