+++ tests/tests_index.js
@@ -1,5 +1,5 @@
 const test = require('ava');
-const m = require('../src');
+const m = require('../src/db');
 
 test('[DYNAMIC] Testing the search on the different website', async t => {
 	await m.get_data();
